Fix ssh client version retrieval
Current implementation of the code which gets the version of SSH client
actually checks the version of SSH server assuming they are equal.

If SSH server is not installed then the code crashes.

This patch does not rely on SSH server to be installed and avoids
throwing an unhandled exception in case SSH client is also not
installed.

Closes-Bug: #1494309